requirements.txt
aiosqlite==0.16.1
click==7.1.2
fastapi==0.63.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.0
iso8601==0.1.14
priority==1.3.0
pydantic==1.8.1
PyPika==0.44.1
pytz==2020.5
starlette==0.13.6
toml==0.10.2
tortoise-orm==0.16.21
typing-extensions==3.7.4.3
unicorn==1.0.2
uvicorn==0.13.4
wsproto==1.0.0
